refactor: extract shared prompts to prompts.py
- Create prompts.py with shared SYSTEM_PROMPT and build_prompt
- Update ollama_service.py and groq_service.py to use shared prompts
- Fix typo: NEVR -> NEVER
